Applied dynamic routing to projects. Updated routing on blogs to allow URL routing with proper props
