--- src/components/ProjectCardDialog/index.js
@@ -1,2 +1,0 @@
-import "./ProjectCardDialog.css";
-export { default } from "./ProjectCardDialog";
