+++ src/pages/ProjectPage/index.js
@@ -1,0 +1,2 @@
+import "./ProjectPage.css";
+export { default } from "./ProjectPage";
